requirements-dev.txt
Faker==26.0.0
pytest==8.2.2
PyYAML==6.0.1
-r requirements.txt
